ec2build.sh: skip the extra download cache for now
Logic needs to be cleaned up and I need to decide on an approach of S3
vs. EBS and how people might be able to use the cache if they aren't
using Amazon Web Services.  The cache needs to be stable around a
release to be sure that we don't lose any sources.

Also updated revision to a version with camera patches.